fix: extract main bodies into headers to clear CodeQL cpp/include-non-header
CodeQL flagged the two \`#include \"../src/X.cpp\"\` directives in the
test fixtures at \`tests/api_server_test.cpp:22\` and
\`tests/program_entry_test.cpp:7\`. The \`#define main X; #include cpp;
#undef main\` trick was a convenient way to reach the main() function
for testing, but it is exactly what \`cpp/include-non-header\` exists
to forbid.

Extract the entire body of each main source into a companion header
that the tests can include directly:

- \`src/main_body.h\` holds \`run_airline_main()\` (inline). The new
  \`src/main.cpp\` is a three-line wrapper that just returns it.
  \`tests/program_entry_test.cpp\` now includes \`main_body.h\` and
  calls \`run_airline_main()\` instead of the renamed main.
- \`src/api_server_main_body.h\` holds every helper, handler,
  register/run/entry function, to_json overloads, and constants from
  the old \`api_server_main.cpp\`, all marked \`inline\`. The new
  \`src/api_server_main.cpp\` is a three-line wrapper that calls
  \`airline_api_server_entry(std::cin, std::cout, std::cerr)\`.
  \`tests/api_server_test.cpp\` and
  \`tests/api_server_entry_and_helpers_cases.inc\` drop the
  \`#define main airline_api_server_entry_main\` renaming — the entry
  tests call \`airline_api_server_entry(...)\` directly which is what
  the renamed main did anyway.

Behavior, coverage, and the Sonar/Codacy/DeepSource gates all stay
unchanged. Only the textual-inclusion pattern is replaced.
