+++ src/main.cpp
@@ -1,17 +1,6 @@
 // cppcheck-suppress-file missingIncludeSystem
-#include "ReservationSystem.h"
-#include <iostream>
+#include "main_body.h"
 
 int main() {
-    std::cout << "Welcome to the Airline Reservation System!" << std::endl;
-    std::cout << "Initializing..." << std::endl;
-
-    ReservationSystem airlineSystem;
-    
-    std::cout << "Initialization Complete. Starting system..." << std::endl;
-    airlineSystem.run();
-
-    std::cout << "Thank you for using the Airline Reservation System." << std::endl;
-
-    return 0;
+    return run_airline_main();
 }
